+++ packages/orq-rc/RELEASES.md
@@ -8249,3 +8249,13 @@ Based on:
 - [typescript v4.8.0-rc.9] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.9 - packages/orq-rc
+
+## 2026-04-13 13:57:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.3 (2.879.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.10] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.10 - packages/orq-rc